Implement single field inherit function
- Move inherit code under inherit dir
- Split out get/send field code from the rest of the inherit code
