+++ server/modules/selva/module/find_index/icb.c
@@ -9,7 +9,6 @@
 #include "selva.h"
 #include "selva_set.h"
 #include "traversal.h"
-#include "traversal_order.h"
 #include "hierarchy.h"
 #include "icb.h"
