+++ server/modules/selva/module/hierarchy/traversal_order.c
@@ -8,7 +8,7 @@
 #include "hierarchy.h"
 #include "selva_lang.h"
 #include "selva_object.h"
-#include "traversal_order.h"
+#include "traversal.h"
 
 /**
  * Used internally to pass the data needed for creating order items.
